+++ CHANGELOG.md
@@ -8,3 +8,4 @@
 * Updated toolbar icons
 * Updated documentation
 * Fixed a bug that prevented thumbnail to be shown
+* Fixed blurry text on some screens with DPI scaling